// RMG/ClientApp/src/app/models/RoleAttribute.ts
export class RoleAttribute {

  employee_Id: string;

  employee_Name: string;

  role_Designation: string;

  role_Projects: string;

  project_Name: string;

  role_Description: string;

  role_Status: string;

  role_StartDate: string;

  role_EndDate: string;

  role_CreatedBy: string;

  role_CreatedDate: Date;

  lastUpdatedBy: string;

  role_LastUpdatedDate: Date;

}